+++ src/sn/EventListenerCtn.ts
@@ -7,9 +7,10 @@
 
 import {BaseTexture, utils} from "pixi.js";
 
-type  IEmitter = BaseTexture
-	| utils.EventEmitter
-	| Window | Document | Element;
+type IEmitter = BaseTexture | utils.EventEmitter | {
+	addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+	removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+}
 
 export class EventListenerCtn {	// リソースリーク対策
 	#vctEvt	: {(): void}[]	= [];